requirements/base.txt
Flask==0.10.1
Flask-RESTful==0.3.3
millipede==1.0
Flask-Redis==0.1.0
